Extract boilerplate code.
